refactoring of teleop and scripts
