+++ configs/controllers/osc_pd/default.yaml
@@ -1,3 +1,5 @@
+# controller: OSC PD Controller
+# Default parameters for the OSC PD controller
 gains.k_pos_x: 600.0
 gains.k_pos_y: 600.0
 gains.k_pos_z: 600.0
@@ -6,8 +8,8 @@ gains.d_pos_x: -1.0
 gains.d_pos_y: -1.0
 gains.d_pos_z: -1.0
 
-gains.k_rot_x: 100.0
-gains.k_rot_y: 100.0
+gains.k_rot_x: 50.0
+gains.k_rot_y: 50.0
 gains.k_rot_z: 50.0
   
 gains.d_rot_x: -1.0
@@ -24,6 +26,6 @@ tool.length: -0.06 # length of the tool attached to the flange, in meters
 log.enabled: true
 log.timing: true
 log.robot_state: false  #  ee position, velocity, joint states, ...
-log.haptic_input: false  # target pose
+log.targets: false  # target pose
 log.virtual_coupling: true  # interface variables and forces
-log.controller_parameters: false  # Gains
+log.controller_parameters: true  # Gains
