+++ experiments/benchmarks/sin_traj.py
@@ -182,8 +182,8 @@ class ExperimentManager:
         """Collect experiment metadata including controller parameters."""
         try:
             # Get controller parameters
-            ctrl_params = robot.haptic_controller_parameters_client.list_parameters()
-            params_values = robot.haptic_controller_parameters_client.get_parameters(ctrl_params)
+            ctrl_params = robot.osc_pd_controller_parameters_client.list_parameters()
+            params_values = robot.osc_pd_controller_parameters_client.get_parameters(ctrl_params)
             controller_parameters = dict(zip(ctrl_params, params_values))
         except Exception as e:
             print(f"Warning: Could not retrieve controller parameters: {e}")
@@ -434,7 +434,7 @@ print(f"🎮 Switching to controller: {CONTROLLER}")
 robot.controller_switcher_client.switch_controller(CONTROLLER)
 
 if CONTROLLER == "osc_pd_controller":
-    robot.haptic_controller_parameters_client.load_param_config(file_path=config_path)
+    robot.osc_pd_controller_parameters_client.load_param_config(file_path=config_path)
     metadata = experiment.collect_controller_metadata(robot, trajectory_params)
 elif CONTROLLER == "cartesian_impedance_controller":
     robot.cartesian_controller_parameters_client.load_param_config(file_path=config_path)
