+++ experiments/probing/simple_move.py
@@ -17,7 +17,7 @@ print(robot.joint_values)
 print(robot.end_effector_wrench)  # added to print initial wrench
 
 robot.controller_switcher_client.switch_controller("osc_pd_controller")
-robot.haptic_controller_parameters_client.load_param_config(
+robot.osc_pd_controller_parameters_client.load_param_config(
     file_path=CONFIG_DIR / "controllers" / "osc_pd" / "default.yaml"
 )
