--- src/arm_client/arm_client/config/control/haptic_controller.yaml
@@ -1,27 +1,0 @@
-# haptic_position_topic: "/haptic_pose"
-# haptic_twist_topic: "/haptic_twist"
-# interface_force_topic: "/fr3/interface_force"
-# interface_torque_topic: "/fr3/interface_torques"
-
-virtual_coupling.k_pos_x: 600.0
-virtual_coupling.k_pos_y: 1000.0
-virtual_coupling.k_pos_z: 1000.0
-
-virtual_coupling.d_pos_x: -1.0
-virtual_coupling.d_pos_y: -1.0
-virtual_coupling.d_pos_z: -1.0
-
-virtual_coupling.k_rot_x: 50.0
-virtual_coupling.k_rot_y: 70.0
-virtual_coupling.k_rot_z: 50.0
-  
-virtual_coupling.d_rot_x: -1.0
-virtual_coupling.d_rot_y: -1.0
-virtual_coupling.d_rot_z: -1.0
-
-log.enabled: true
-log.timing: true
-log.robot_state: true
-log.haptic_input: true
-log.virtual_coupling: false
-log.controller_parameters: false
