+++ src/arm_client/scripts/02_move_to.py
@@ -19,8 +19,11 @@ robot.controller_switcher_client.switch_controller("osc_pd_controller")
 #     # file_path="config/control/clipped_cartesian_impedance.yaml"
 #     file_path=CONFIG_DIR / "control" / "default_cartesian_impedance.yaml"
 # )
+robot.osc_pd_controller_parameters_client.load_param_config(
+    file_path=CONFIG_DIR / "controllers" / "osc_pd" / "default.yaml"
+)
 
-position = [0.4, -0.2, 0.4]  # x, y, z in meters
+position = [0.4, 0.2, 0.3]  # x, y, z in meters
 robot.move_to(position=position, speed=0.05)
 
 print("Done")
