+++ src/arm_client/scripts/03_figure_eight.py
@@ -38,9 +38,9 @@ max_time = 8.0
 #     file_path=CONFIG_DIR / "controllers" / "default_cartesian_impedance.yaml"
 # )
 
-robot.controller_switcher_client.switch_controller("haptic_controller")
-robot.haptic_controller_parameters_client.load_param_config(
-    file_path=CONFIG_DIR / "controllers" / "haptic_controller.yaml"
+robot.controller_switcher_client.switch_controller("osc_pd_controller")
+robot.osc_pd_controller_parameters_client.load_param_config(
+    file_path=CONFIG_DIR / "controllers" / "osc_pd" / "default.yaml"
 )
