+++ src/robot_arm/franka_rim/scripts/experiment_setup_delay_rim.py
@@ -19,7 +19,7 @@ robot.wait_until_ready(timeout=2.0)
 # --- To Switch Controllers ---
 controller_name = "osc_pd_controller"  # osc_pd_controller, gravity_compensation, ....
 robot.controller_switcher_client.switch_controller(controller_name)
-robot.haptic_controller_parameters_client.load_param_config(
+robot.osc_pd_controller_parameters_client.load_param_config(
     file_path=CONFIG_DIR / "controllers" / "osc_pd" / "default.yaml"
 )
