+++ src/teleop/teleop/setup.py
@@ -30,9 +30,9 @@ setup(
     entry_points={
         "console_scripts": [
             "joy_teleop = teleop.joy_teleop:main",
-            "i3_teleop = teleop.i3_teleop:main",
+            "i3_teleop_old = teleop.i3_teleop_old:main",
             "i3_teleop_position = teleop.i3_teleop_position:main",
-            "i3_teleop_virtual_coupling = teleop.i3_teleop_virtual_coupling:main",
+            "i3_teleop = teleop.i3_teleop:main",
         ],
     },
 )
